fix: Refactor InsigthsLayoutHeader
