+++ src/layouts/InsightsLayout/Header.vue
@@ -8,27 +8,25 @@
     <h1 class="header__title">
       {{ isInHome ? 'Insights' : selectedDashboardLabel }}
     </h1>
-    <section class="header__filters">
+    <section class="header__content">
       <UnnnicSelectSmart
         v-model="selectedDashboard"
         :options="dashboards"
         orderedByIndex
       />
-      <UnnnicInputDatePicker
-        class="filters__date-picker"
-        v-model="filterDate"
-        size="sm"
-        inputFormat="DD/MM/YYYY"
-      />
+
+      <InsightsLayoutHeaderFilters />
     </section>
   </header>
 </template>
 
 <script>
-import moment from 'moment';
+import InsightsLayoutHeaderFilters from './HeaderFilters.vue';
 
 export default {
   name: 'InsightsLayoutHeader',
+
+  components: { InsightsLayoutHeaderFilters },
 
   data() {
     return {
@@ -42,10 +40,6 @@ export default {
         { value: 'human-service', label: 'Atendimento Humano' },
       ],
       selectedDashboard: [],
-      filterDate: {
-        start: moment().subtract(1, 'day').format('YYYY-MM-DD'),
-        end: moment().format('YYYY-MM-DD'),
-      },
     };
   },
 
@@ -86,25 +80,6 @@ export default {
 
       this.selectedDashboard[0] = dashboardRelativeToPath;
     },
-
-    routeUpdateFilterDate() {
-      const { query } = this.$route;
-
-      if (query.startDate && query.endDate) {
-        this.filterDate = {
-          start: query.startDate,
-          end: query.endDate,
-        };
-      }
-    },
-
-    retainRouteQueries(newRoute, oldRoute) {
-      const oldQueryKeys = Object.keys(oldRoute?.query);
-
-      if (oldQueryKeys.length) {
-        this.$router.replace({ name: newRoute.name, query: oldRoute.query });
-      }
-    },
   },
 
   watch: {
@@ -115,16 +90,8 @@ export default {
     },
     $route(newRoute, oldRoute) {
       if (newRoute.name !== oldRoute.name) {
-        this.retainRouteQueries(newRoute, oldRoute);
         this.routeUpdateSelectedDashboard();
-        this.routeUpdateFilterDate();
       }
-    },
-    filterDate(newDate) {
-      this.$router.replace({
-        name: this.$route.name,
-        query: { startDate: newDate.start, endDate: newDate.end },
-      });
     },
   },
 };
@@ -142,7 +109,7 @@ export default {
     font-family: $unnnic-font-family-primary;
     line-height: $unnnic-line-height-large * 2;
   }
-  .header__filters {
+  .header__content {
     display: flex;
     flex-direction: row;
     gap: $unnnic-spacing-xs;
@@ -150,24 +117,6 @@ export default {
     :deep(.unnnic-select-smart) {
       .dropdown-data {
         z-index: 2;
-      }
-    }
-
-    .filters__date-picker {
-      display: grid;
-
-      :deep(.unnnic-form-input) {
-        height: 100%;
-
-        .unnnic-icon {
-          // It was necessary to follow bad practices here (px) because
-          // of how the component was initially implemented.
-          top: $unnnic-spacing-sm - 1px;
-        }
-
-        .input {
-          font-size: $unnnic-font-size-body-gt;
-        }
       }
     }
   }
